11.11.2023

+++ api/socket/socket/ping-pong.ts
@@ -1,27 +1,48 @@
 import WebSocket from "ws";
 import { IWebSocket } from "../server";
-import { inject } from "inversify";
+import { inject, injectable } from "inversify";
 import { TYPES } from "../../../types";
 import { Rooms } from "./rooms";
 import { PointerService } from "../../../services/pointer.service";
+import { ZoneService } from "../../../services/zone.service";
+import { WeaponService } from "../../../services/weapon.service";
 
+@injectable()
 export class PingPong {
 
     @inject(TYPES.Rooms) private _rooms!: Rooms
+    @inject(TYPES.ZoneService) private _zoneService!: ZoneService
+    @inject(TYPES.WeaponService) private _weaponService!: WeaponService
     @inject(TYPES.PointerService) private _pointerService!: PointerService
 
-    private _wss: WebSocket.Server
+    // @inject(TYPES.VkUserRepository) private _vkUserRepository!: VkUserRepository
+    private _wss: WebSocket.Server | null = null
 
-    constructor(wss: WebSocket.Server) {
-        this._wss = wss
-
-        this.start()
+    constructor() {
     }
 
     public async deleteUser(userId: number) {
+
+        if (!userId) return
+
+        console.log('deleteUser')
+
         const _pointer = await this._pointerService.memoryGetById(userId)
+        _pointer.areal = 0
+        const _zone = await this._zoneService.memoryGetById(userId)
+        const _weapon = await this._weaponService.memoryGetById(_pointer.weapons[0])
+
+        console.log('//////////////// /////////////deleteUser _pointer.pos', _pointer.pos)
+
+        await this._pointerService.baseUpdate(_pointer)
+        await this._zoneService.baseUpdate(_zone)
+        await this._weaponService.baseUpdate(_weapon)
+
         this._rooms.areals.deleteClient(_pointer.zoneId, _pointer.areal)
-        await this._pointerService.baseInsert(_pointer)
+
+        await this._weaponService.remove(_pointer.weapons[0])
+        this._pointerService.remove(userId)
+        this._zoneService.remove(userId)
     }
 
     async each(ws: IWebSocket) {
@@ -39,11 +60,14 @@ export class PingPong {
     }
 
     pingPong() {
-        // const context = this
-        this._wss.clients.forEach(this.each.bind(this));
+        console.log('pingPong')
+        if (this._wss)
+            // const context = this
+            this._wss.clients.forEach(this.each.bind(this));
     }
 
-    start() {
+    start(wss: WebSocket.Server) {
+        this._wss = wss
         // const context = this
         setInterval(this.pingPong.bind(this), 10000);
     }
